requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
decorator==4.4.2
discord.py==1.3.3
idna==2.10
idna-ssl==1.1.0
multidict==4.7.6
pycparser==2.20
python-telegram-bot==12.8
six==1.15.0
tornado==6.0.4
typing-extensions==3.7.4.2
websockets==8.1
yarl==1.4.2
